requirements.txt
Flask>=0.10.1
leancloud-sdk
pyopenssl
ndg-httpsclient
pyasn1
nose
